add images and links to vericomp consortium members
Note that these are taken from the robmosys pages but not sure if
they make sense. Uni-bielefeld has the old logo, cor-lab goes to
somewhere that is not cor-lab, TU-Braunschweig is mentioned (instead
of the IRP), etc. I kept them as is for sake of consistency. Not
that it matters anyway...